Refine package split: rename pypromice-ingest → pypromice-tx, move file IO to pipeline
- Rename packages/pypromice-ingest/ → packages/pypromice-tx/
  - Module path: pypromice.ingest.* → pypromice.tx.*
  - Scope narrowed to transmission/decoding only (email, Iridium SBD, payload)
  - Updated imports in get_l0tx.py, get_msg.py, rest_api_client.py
  - Updated error message in rest_api_client.py (pypromice-ingest → pypromice-tx)

- Move file IO modules to pypromice-pipeline (pypromice.ingest.* namespace)
  - git.py, l0.py, l0_repository.py, toa5.py from ingest → pipeline/src/pypromice/ingest/
  - No import changes needed (pypromice.ingest.* path preserved)
  - pypromice-pipeline no longer depends on pypromice-tx

- Clean up root tests/unit/ (all unit tests now co-located with packages)
  - Remove tests/unit/ and tests/utilities.py from repo root
  - Root tests/ now contains only e2e integration tests

- Update pyproject.toml, publish workflow, and unit_test.yml accordingly
