+++ packages/pypromice-pipeline/src/pypromice/ingest/git.py
