+++ packages/pypromice-tx/README.md
@@ -1,0 +1,17 @@
+# pypromice-tx
+
+PROMICE/GC-Net AWS data transmission package: IMAP email fetching, Iridium SBD payload decoding, and L0 message handling.
+
+Part of the [pypromice](https://github.com/GEUS-Glaciology-and-Climate/pypromice) monorepo.
+
+## Installation
+
+```bash
+pip install pypromice-tx
+```
+
+For Gmail REST API support:
+
+```bash
+pip install pypromice-tx[google]
+```
