--- tests/unit/bufr_export/__init__.py
@@ -1,7 +1,0 @@
-import pytest
-
-pytest.importorskip(
-    "eccodes",
-    reason="eccodes not installed. BUFR export tests skipped. "
-           "Install with `pip install pypromice[eccodes]`."
-)
